Add rate limiter, authentication mocks, and user role management
- Implement Fixed Window Rate Limiter with Allow method and reset logic.
- Create tests for Fixed Window Rate Limiter to validate request handling and concurrency.
- Introduce Limiter interface and mock implementations for testing.
- Add Redis-based caching for user and role management with appropriate methods.
- Create mocks for user and role stores to facilitate testing.
- Develop CreatePostForm component for posting content to the API.

+++ cmd/migrate/migrations/000014_add_role_to_users.up.sql
@@ -1,5 +1,5 @@
 ALTER TABLE IF EXISTS users
-ADD COLUMN IF NOT EXISTS role_id int REFERENCES roles(id) DEFAULT 1;
+ADD COLUMN IF NOT EXISTS role_id BIGINT REFERENCES roles(id) DEFAULT 1;
 UPDATE users
 SET role_id = (
         SELECT id
